requirements.txt
python-dotenv
torch
transformers
streamlit







